Move the flashcard-suggestion reads into queries/
The three suggestion endpoints are the ADR's halfway option at its most
minimal: they relocate as-is, keep asking repositories for aggregates and
keep returning FlashcardSuggestion. Nothing here would gain from a query
adapter — each is one find_by_id plus an AI call — but the package
placement is what lets use_cases/ finish as commands-only.

FlashcardSuggestion moves with them, so learning/use_cases/dtos/ is gone.
