--- backend/src/application/learning/use_cases/dtos/__init__.py
@@ -1,5 +1,0 @@
-"""DTOs for learning use cases."""
-
-from src.application.learning.use_cases.dtos.flashcard_ai_dtos import FlashcardSuggestion
-
-__all__ = ["FlashcardSuggestion"]
