--- backend/src/application/learning/use_cases/dtos/flashcard_ai_dtos.py
@@ -1,11 +1,0 @@
-"""DTOs for flashcard AI use cases."""
-
-from dataclasses import dataclass
-
-
-@dataclass
-class FlashcardSuggestion:
-    """Simple data class for AI suggestions."""
-
-    question: str
-    answer: str
