+++ backend/src/infrastructure/learning/routers/ai_note_flashcard_suggestions.py
@@ -5,7 +5,7 @@ from typing import Annotated
 import structlog
 from fastapi import APIRouter, Depends, status
 
-from src.application.learning.use_cases.flashcards.get_note_flashcard_suggestions_use_case import (
+from src.application.learning.queries.get_note_flashcard_suggestions_use_case import (
     GetNoteFlashcardSuggestionsUseCase,
 )
 from src.core import container
